Change array type to Double and update input method in array sum program; add skeleton for Question2 class

+++ chapter6/Question1.java
@@ -3,11 +3,11 @@ import java.util.Scanner;
 public class Question1 {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int[] arr = new int[5];
-        int sum = 0;
+        Double[] arr = new Double[5];
+        double sum = 0;
         for (int i = 0; i < arr.length; i++) {
             System.out.print("Enter element " + (i + 1) + ": ");
-            arr[i] = sc.nextInt();
+            arr[i] = sc.nextDouble();
             sum += arr[i];
         }
